requirements.txt
alembic==1.0.0
appdirs==1.4.3
blinker==1.4
boto3==1.7.83
botocore==1.10.83
certifi==2018.8.13
chardet==3.0.4
click==6.7
docutils==0.14
Faker==0.8.17
Flask==1.0.2
Flask-Assets==0.12
Flask-Compress==1.4.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.2.1
Flask-RQ==0.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-SSLify==0.1.5
Flask-WTF==0.14.2
gunicorn==19.9.0
honcho==1.0.1
html2text==2018.1.9
humanfriendly==4.16
idna==2.7
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
jsmin==2.2.2
jsonpickle==0.9.6
lxml==4.2.4
Mako==1.0.7
MarkupSafe==1.0
packaging==17.1
psycopg2==2.7.5
pur==5.0.0
pycurl==7.43.0.2
PyJWT==1.6.4
pyparsing==2.2.0
python-dateutil==2.7.3
python-docx==0.8.6
python-editor==1.0.3
raygun4py==3.1.6
redis==2.10.6
requests==2.19.1
rq==0.12.0
s3transfer==0.1.13
six==1.11.0
SQLAlchemy==1.2.10
text-unidecode==1.2
urllib3==1.23
webassets==0.12.1
Werkzeug==0.14.1
Wikipedia-API==0.3.7
wptools==0.4.16
WTForms==2.2.1
